Refactor: features 폴더 대분류
