+++ src/features/market/ui/MarketIndexSummary.tsx
