+++ src/features/stock/investmentStatus/ui/InvestmentStatus.tsx
