+++ src/features/stock/stockChartBoard/ui/LiveStockTable.tsx
