+++ src/features/stock/stockChartBoard/ui/StockChart.tsx
