+++ src/features/stock/stockChartBoard/ui/StockChartBoard.tsx
