+++ src/features/stock/stockDetail/ui/StockDetail.tsx
