+++ src/features/user/myInvestment/ui/InvestmentInfo.tsx
