+++ src/features/user/profitGraph/ui/ProfitGraph.tsx
