+++ src/features/user/realizedProfit/ui/ProfitDetail.tsx
