+++ src/pages/OverviewPage.tsx
@@ -1,7 +1,7 @@
 import InvestmentStatus from "@/features/investmentStatus/ui/InvestmentStatus";
-import MyAmount from "@/features/myAmount/ui/MyAmount";
 import ProfitGraph from "@/features/profitGraph/ui/ProfitGraph";
-import UserInfo from "@/features/userInfo/ui/UserInfo";
+import MyAmount from "@/features/user/myAmount/ui/MyAmount";
+import UserInfo from "@/features/user/userInfo/ui/UserInfo";
 import Footer from "@/shared/components/molecules/Footer";
 import Header from "@/shared/components/molecules/Header";
 import ScreenTooSmall from "@/shared/components/organisms/ScreenTooSmall";
